apm: make http.client_ip an appsec-only feature
