+++ contrib/internal/httptrace/config.go
@@ -20,10 +20,6 @@ const (
 	envQueryStringDisabled = "DD_TRACE_HTTP_URL_QUERY_STRING_DISABLED"
 	// envQueryStringRegexp is the name of the env var used to specify the regexp to use for query string obfuscation.
 	envQueryStringRegexp = "DD_TRACE_OBFUSCATION_QUERY_STRING_REGEXP"
-	// envClientIPHeader is the name of the env var used to specify the IP header to be used for client IP collection.
-	envClientIPHeader = "DD_TRACE_CLIENT_IP_HEADER"
-	// envClientIPHeader is the name of the env var used to disable client IP tag collection.
-	envClientIPHeaderDisabled = "DD_TRACE_CLIENT_IP_HEADER_DISABLED"
 )
 
 // defaultQueryStringRegexp is the regexp used for query string obfuscation if `envQueryStringRegexp` is empty.
@@ -31,15 +27,11 @@ var defaultQueryStringRegexp = regexp.MustCompile("(?i)(?:p(?:ass)?w(?:or)?d|pas
 
 type config struct {
 	queryStringRegexp *regexp.Regexp // specifies the regexp to use for query string obfuscation.
-	clientIPHeader    string         // specifies the header to use for IP extraction if client IP tag collection is enabled.
-	clientIP          bool           // reports whether the IP should be extracted from the request headers and added to span tags.
 	queryString       bool           // reports whether the query string should be included in the URL span tag.
 }
 
 func newConfig() config {
 	c := config{
-		clientIPHeader:    os.Getenv(envClientIPHeader),
-		clientIP:          !internal.BoolEnv(envClientIPHeaderDisabled, false),
 		queryString:       !internal.BoolEnv(envQueryStringDisabled, false),
 		queryStringRegexp: defaultQueryStringRegexp,
 	}
